Refactor: se mueven archivos y se corrige el Spanglish en archivos
