+++ src/pages/AdminPages/UsersDashboard.jsx
@@ -3,7 +3,7 @@ import UserTable from "../../components/Tables/UserTable";
 import { Link } from "react-router-dom";
 import { ChevronRight } from "lucide-react";
 
-const UserPanel = () => {
+const UsersDashboard = () => {
   return (
     <>
       <div className="container mx-auto my-20">
@@ -16,4 +16,4 @@ const UserPanel = () => {
   );
 };
 
-export default UserPanel;
+export default UsersDashboard;
