+++ src/pages/AdminPages/VacanciesDashboard.jsx
@@ -1,17 +1,17 @@
 import React from "react";
-import VacantesTable from "../../components/Tables/VacantesTable";
+import VacanciesTable from "../../components/Tables/VacanciesTable";
 import { Link } from "react-router-dom";
 import { ChevronRight } from "lucide-react";
 
-const VacantesPanel = () => {
+const VacanciesDashboard = () => {
   return (
     <div className="container mx-auto my-20">
       <Link to="/admin" className="mt-5 text-blue-500 flex">
         Volver atras <ChevronRight />
       </Link>
-      <VacantesTable />
+      <VacanciesTable />
     </div>
   );
 };
 
-export default VacantesPanel;
+export default VacanciesDashboard;
