+++ src/routes/RoutesViews.jsx
@@ -4,14 +4,14 @@ import { useLoginFirebase } from "../hooks/useLoginFirebase";
 import HomePage from "../pages/HomePage";
 import LoginPage from "../pages/Auth/LoginPage";
 import App404 from "../pages/App404";
-import UserPanel from "../pages/AdminPages/UserPanel";
 import UserRegister from "../pages/Auth/UserRegister";
 import AdminView from "../pages/AdminPages/AdminView";
 import ProtectedRoute from "./ProtectedRoute";
 import Navbar from "../components/Navbar";
-import VacantesPanel from "../pages/AdminPages/VacantesPanel";
-import Vacancies from "../pages/vacancies";
-import VacancyGallery from "../pages/PublicPages/VacancyGallery";
+import VacanciesDashboard from "../pages/AdminPages/VacanciesDashboard";
+import UsersDashboard from "../pages/AdminPages/UsersDashboard";
+import CreateVacancies from "../pages/AdminPages/CreateVacancies";
+import VacanciesGallery from "../pages/PublicPages/VacanciesGallery";
 
 const RoutesViews = () => {
   const [isAuthenticated, setIsAuthenticated] = useState(
@@ -38,8 +38,7 @@ const RoutesViews = () => {
       <Routes>
         {/* Rutas Públicas */}
         <Route path="/" element={<HomePage />} />
-        <Route path="/ver/vacantes" element={<VacancyGallery />} />
-        <Route path="/vacantes" element={<Vacancies />} />
+        <Route path="/ver/vacantes" element={<VacanciesGallery />} />
         <Route path="/login" element={<LoginPage login={login} />} />
 
         {/* Rutas Protegidas */}
@@ -55,7 +54,7 @@ const RoutesViews = () => {
           path="/admin/panelUsuarios"
           element={
             <ProtectedRoute isAuthenticated={isAuthenticated}>
-              <UserPanel />
+              <UsersDashboard />
             </ProtectedRoute>
           }
         />
@@ -63,15 +62,23 @@ const RoutesViews = () => {
           path="/admin/panelVacantes"
           element={
             <ProtectedRoute isAuthenticated={isAuthenticated}>
-              <VacantesPanel />
+              <VacanciesDashboard />
             </ProtectedRoute>
           }
         />
         <Route
-          path="/admin/crearUsuario"
+          path="/admin/crear/usuario"
           element={
             <ProtectedRoute isAuthenticated={isAuthenticated}>
               <UserRegister />
+            </ProtectedRoute>
+          }
+        />
+        <Route
+          path="/crear/vacante"
+          element={
+            <ProtectedRoute isAuthenticated={isAuthenticated}>
+              <CreateVacancies />
             </ProtectedRoute>
           }
         />
